refactor: userProgress
