+++ apps/client/pages/index.vue
@@ -77,49 +77,15 @@
 </template>
 
 <script setup lang="ts">
-import { fetchUpdateProgress, fetchUserProgress } from "~/api/userProgress";
 import { useUserStore } from "~/store/user";
 import { registerShortcut, cancelShortcut } from "~/utils/keyboardShortcuts";
-
-
-const useProgress = () => {
-
-  const activeCourseId = ref(1)
-  const ACTIVE_COURSE_ID = 'activeCourseId'
-
-  const initing = ref(false)
-  const initProgress = async () => {
-    initing.value = true
-    const { courseId } = await fetchUserProgress()
-    activeCourseId.value = +courseId
-    updateProgressLocal(+courseId)
-    initing.value = false
-  }
-
-  const updateProgress = async (courseId: number) => {
-    const { courseId: updatedCourseId } = await fetchUpdateProgress({courseId}) 
-    updateProgressLocal(updatedCourseId)
-  }
-
-  const updateProgressLocal = async (courseId: number) => {
-    localStorage.setItem(ACTIVE_COURSE_ID, `${courseId}`)
-  }
-  
-  return {
-    activeCourseId,
-    initing,
-    updateProgressLocal,
-    updateProgress,
-    initProgress
-  }
-}
 
 const { handleKeydown, initing } = useShortcutToGame();
 
 function useShortcutToGame() {
   const router = useRouter();
   const userStore = useUserStore();
-  const { activeCourseId, initing, initProgress } = useProgress()
+  const { activeCourseId, initing, initProgress } = useUserProgress()
 
   function handleKeydown() {
     if (userStore.user) {
